Refactor server and app for testability
Moving the instantiation of the server away from the application itself
allows us to import the express app as a moodule and test it. This keeps
us from actually spinning up a server instance on every import.
